raft-requirements.txt
requests-oauthlib-2. 
markdown-3.4.1       
fonttools-4.51.0     
importlib-metadata-7 
protobuf-4.25.3      
libgfortran-ng-11.2. 
libcups-2.4.2        
yarl-1.11.0          
libunistring-0.9.10  
pooch-1.7.0          
lame-3.100           
ffmpeg-4.2.2         
libssh2-1.11.0       
brotli-bin-1.0.9     
pyqt5-sip-12.13.0    
pyparsing-3.0.9      
pyopenssl-24.2.1     
openjpeg-2.5.2       
ninja-1.10.2         
wheel-0.44.0         
pillow-10.4.0        
python-dateutil-2.9. 
google-auth-oauthlib 
setuptools-75.1.0    
lerc-3.0             
aiohappyeyeballs-2.4 
opencv-4.10.0        
libtiff-4.5.1        
brotli-1.0.9         
numpy-base-1.24.3    
libbrotlienc-1.0.9   
pytorch-1.6.0        
sip-6.7.12           
requests-2.32.3      
mkl_random-1.2.4     
python-3.8.19        
mysql-5.7.24         
attrs-23.1.0         
gstreamer-1.14.1     
tensorboard-data-ser 
importlib_resources- 
x264-1!157.20191217  
gst-plugins-base-1.1 
platformdirs-3.10.0  
markupsafe-2.1.3     
libclang13-14.0.6    
libidn2-2.3.4        
contourpy-1.0.5      
mkl_fft-1.3.8        
ply-3.11             
matplotlib-3.7.2     
matplotlib-base-3.7. 
libvpx-1.7.0         
werkzeug-3.0.3       
cryptography-43.0.0  
pysocks-1.7.1        
libtasn1-4.19.0      
qt-main-5.15.2       
jpeg-9e              
libdeflate-1.17      
click-8.1.7          
cffi-1.17.1          
hdf5-1.12.1          
cyrus-sasl-2.1.28    
libllvm14-14.0.6     
libbrotlidec-1.0.9   
frozenlist-1.4.0     
brotli-python-1.0.9  
libgfortran5-11.2.0  
idna-3.7             
tornado-6.4.1        
cycler-0.11.0        
packaging-24.1       
pyqt-5.15.10         
pyjwt-2.8.0          
libopus-1.3.1        
scipy-1.10.1         
urllib3-2.2.2        
oauthlib-3.2.2       
openh264-2.1.1       
numpy-1.24.3         
libbrotlicommon-1.0. 
certifi-2024.8.30    
aiohttp-3.10.5       
graphite2-1.3.14     
tensorboard-plugin-w 
tomli-2.0.1          
unicodedata2-15.1.0  
libclang-14.0.6      
absl-py-2.1.0        
mkl-service-2.4.0    
gnutls-3.6.15        
cairo-1.16.0         
zipp-3.17.0          
libnghttp2-1.57.0    
libpq-12.17          
grpcio-1.62.2        
blinker-1.6.2        
cudatoolkit-10.1.243 
tensorboard-2.12.1   
eigen-3.4.0          
pip-24.2             
harfbuzz-4.3.0       
libwebp-base-1.3.2   
libcurl-8.9.1        
kiwisolver-1.4.4     
nettle-3.7.3         
torchvision-0.7.0    
async-timeout-4.0.3  
cachetools-5.3.3     
multidict-6.0.4      
pixman-0.40.0        
ninja-base-1.10.2    
google-auth-2.29.0   